Bug fixed

+++ trunk/src/misc.h
@@ -30,7 +30,7 @@ inline double get_time_interval(int fps)
 {
   
   if(!fps)return 0;
-  return 1000.0/((float)fps);
+  return 1000.0/((double)fps);
   
 }
